# swagger.yaml
swagger: "2.0"
info:
  description: ""
  version: "1.0.0"
  title: "Micro Service Project - Coalition"
host: "coalition.com"
basePath: "/api/"
...
schemes:
  - "https"
  - "http"
paths:
...
  /servers/members:
    post:
      tags:
        - "Server/Member"
      summary : "Add user to server"
      produces :
        - "application/json"
      parameters:
        - in: body
          name : body
          required: true
          schema:
            type: object
            properties: 
              serverID:
                type: integer
              userID:
                type: integer
      responses:
        "201":
          description: "successful operation"
        "404":
          description: "User or Server not found"
        "500":
          description: "Internal server error"
    get:
      tags:
        - "Server/Member"
      summary: "Get all members of all servers"
      produces:
        - "application/json"
      responses:
        "200":
          description:  "successful operation"
          schema:
            type : array
            items:
              type: object
              properties:
                server_id: 
                  type : integer
                user_id :
                  type : integer
                created_at :
                  type : string
        "500":
          description: "Internal server error"
    delete:
      tags:
        - "Server/Member"
      summary : "Delete member from server"
      produces :
        - "application/json"
      parameters:
        - in: body
          name : body
          required: true
          schema:
            type: object
            properties: 
              serverID:
                type: integer
              userID:
                type: integer
      responses:
        "204":
          description: "successful operation"
        "404":
          description: "Server or User not found / User is not member"
        "500":
          description: "Internal server error"
  /servers/members/server/{serverID} :
    get :
      tags:
      - "Server/Member"
      summary: "Get all members of a server"
      produces:
        - "application/json"
      parameters:
      - name: "serverID"
        in: path
        description: "ID of server"
        required: true
        type: integer
      responses:
        "200":
          description:  "successful operation"
          schema:
            type : array
            items:
              type: object
              properties:
                user_id :
                  type : integer
                created_at :
                  type : string
        "404":
          description: "Server not found"
        "500":
          description: "Internal server error"
  /servers/members/user/{userID} :
    get :
      tags:
      - "Server/Member"
      summary: "Get all servers of a user"
      produces:
        - "application/json"
      parameters:
      - in: path
        name: "userID"
        description: "ID of user"
        required: true
        type: integer
      responses:
        "200":
          description:  "successful operation"
          schema:
            type : array
            items:
              type: object
              properties:
                server_id :
                  type : integer
                created_at :
                  type : string
        "404":
          description: "User not found"
        "500":
          description: "Internal server error"
  /servers/channels/:
    post:
      tags:
        - "Server/Channel"
      summary : "Add channel to server"
      produces :
        - "application/json"
      parameters:
        - in: body
          name : body
          required: true
          schema:
            type: object
            properties: 
              serverID:
                type: integer
              userID:
                type: integer
              channelName:
                type: string
      responses:
        "201":
          description: "successful operation"
        "403":
          description: "Permission denied"
        "404":
          description: "User or Server not found"
        "500":
          description: "Internal server error"
    get:
      tags:
        - "Server/Channel"
      summary: "Get all channels of all servers"
      produces:
        - "application/json"
      responses:
        "200":
          description:  "successful operation"
          schema:
            type : array
            items:
              type: object
              properties:
                server_id :
                  type: integer
                conv_id :
                  type: integer
                channel_name :
                  type: string
                created_at :
                  type: string
                updated_at :
                  type: string
        "500":
          description: "Internal server error"
    delete:
      tags:
        - "Server/Channel"
      summary : "Delete channel from server"
      produces :
        - "application/json"
      parameters:
        - in: body
          name : body
          required: true
          schema:
            type: object
            properties: 
              serverID:
                type: integer
              userID:
                type: integer
              convID:
                type: integer
      responses:
        "204":
          description: "successful operation"
        "403":
          description: "Permission denied"
        "404":
          description: "Server, User or Channel not found"
        "500":
          description: "Internal server error"
  /servers/channels/server/{serverID} :
    get :
      tags:
      - "Server/Channel"
      summary: "Get all channels of a server"
      produces:
        - "application/json"
      parameters:
      - in: path
        name: "serverID"
        description: "ID of server"
        required: true
        type: integer
      - in : body
        name : body
        required: true
        schema:
          type: object
          properties: 
            userID:
              type: integer
      responses:
        "200":
          description:  "successful operation"
          schema:
            type : array
            items:
              type: object
              properties:
                conv_id :
                  type: integer
                channel_name :
                  type: string
                created_at :
                  type: string
                updated_at :
                  type: string
        "403":
          description: "Permission denied"
        "404":
          description: "Server or User not found"
        "500":
          description: "Internal server error"
...
  /groups/members:
    post:
      tags:
        - "Group/Member"
      summary : "Add user to group"
      produces :
        - "application/json"
      parameters:
        - in: body
          name : body
          required: true
          schema:
            type: object
            properties: 
              groupID:
                type: integer
              userID:
                type: integer
      responses:
        "201":
          description: "successful operation"
        "404":
          description: "User or Group not found"
        "500":
          description: "Internal server error"
    get:
      tags:
        - "Group/Member"
      summary: "Get all members of all groups"
      produces:
        - "application/json"
      responses:
        "200":
          description:  "successful operation"
          schema:
            type : array
            items:
              type: object
              properties:
                group_id: 
                  type : integer
                user_id :
                  type : integer
                created_at :
                  type : string
        "500":
          description: "Internal server error"
    delete:
      tags:
        - "Group/Member"
      summary : "Delete member from group"
      produces :
        - "application/json"
      parameters:
        - in: body
          name : body
          required: true
          schema:
            type: object
            properties: 
              groupID:
                type: integer
              userID:
                type: integer
      responses:
        "204":
          description: "successful operation"
        "404":
          description: "Group or User not found / User is not member"
        "500":
          description: "Internal server error"
...
  /groups/members/user/{userID} :
    get :
      tags:
        - "Group/Member"
      summary: "Get all groups of a user"
      produces:
        - "application/json"
      parameters:
        - in: path
          name: "userID"
          description: "ID of user"
          required: true
          type: integer
      responses:
        "200":
          description:  "successful operation"
          schema:
            type : array
            items:
              type: object
              properties:
                group_id :
                  type : integer
                created_at :
                  type : string
        "404":
          description: "User not found"
        "500":
          description: "Internal server error"
...